checkpatch: Ignore SPDX_LICENSE_TAG
The Linux kernel expects the SPDX license tag in the first line of each
source code file in a comment.

In the context of the Linux kernel repository this makes sense because
they have many different license headers across their codebase. Moving
the tag to the first line of the source code files makes it easier for
analyzers to see the license of each file.

In the Trusted Firmware, we control all headers and make sure that they
follow the same pattern, so this is not needed.

Change-Id: Ie19802c7b65b1bdd63da9ece64311aec1f8ad7fe
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>